UpdateUploadImagesInController

+++ CinemaTicketBooking/Areas/Admin/Controllers/ActorsController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using System.Numerics;
+using Microsoft.AspNetCore.Mvc;
 
 namespace CinemaTicketBooking.Areas.Admin.Controllers
 {
@@ -44,28 +45,26 @@ namespace CinemaTicketBooking.Areas.Admin.Controllers
         [HttpPost]
         public IActionResult Create(ActorsCreateVM actor, IFormFile img)
         {
-            ModelState.Remove("Movie");
+            ModelState.Remove("Img");
+            ModelState.Remove("Movies");
             if (!ModelState.IsValid)
             {
                 actor.Movies = _context.Movies.AsNoTracking().AsQueryable();
                 return View(actor);
             }
-            if (img is not null && img.Length > 0)
-            {
                 var newFileName = Guid.NewGuid().ToString() + DateTime.UtcNow.ToString("yyyy-MM-dd") + Path.GetExtension(img.FileName);
                 var filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\img\\actors_images", newFileName);
                 using (var stream = System.IO.File.Create(filePath))
                 {
                     img.CopyTo(stream);
                 }
-                actor.Img = newFileName;
-            }
+            
             var actors = new Actors
             {
                 Name = actor.Name,
                 Description = actor.Description,
+                Img = newFileName,
                 MovieId = actor.MovieId,
-                Img = actor.Img
             };
             _context.Actors.Add(actors);
             _context.SaveChanges();
@@ -74,18 +73,32 @@ namespace CinemaTicketBooking.Areas.Admin.Controllers
         [HttpGet]
         public IActionResult Edit([FromRoute] int id)
         {
+
             var actor = _context.Actors.Find(id);
+            var movies = _context.Movies.AsNoTracking().AsQueryable();
             if (actor is null)
                 return NotFound();
-            return View(actor);
+            return View(new ActorsUpdateREsponseVM
+            { 
+Name = actor.Name,
+Description = actor.Description,
+Img = actor.Img,
+MovieId = actor.MovieId,
+                Movies = movies.AsEnumerable()
+            });
         }
         [HttpPost]
-        public IActionResult Edit(Actors actor, IFormFile? img)
+        public IActionResult Edit(ActorsUpdateREsponseVM actorVM, IFormFile? img)
         {
-            ModelState.Remove("Movie");
+            ModelState.Remove("Actors.Movie");
+            ModelState.Remove("Img");
+            ModelState.Remove("Movies");
             if (!ModelState.IsValid)
-                return View(actor);
-            Actors? existingActor = _context.Actors.AsNoTracking().FirstOrDefault(b => b.Id == actor.Id);
+            { 
+                actorVM.Movies = _context.Movies;
+                return View(actorVM);
+            }
+            var existingActor = _context.Actors.FirstOrDefault(b => b.Id == actorVM.Id);
             if (existingActor is null)
                 return NotFound();
             if (img is not null && img.Length > 0)
@@ -96,18 +109,21 @@ namespace CinemaTicketBooking.Areas.Admin.Controllers
                 {
                     img.CopyTo(stream);
                 }
-                var oldFilePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\img\\actors_images", existingActor.Img);
-                if (System.IO.File.Exists(oldFilePath))
+                if (!string.IsNullOrEmpty(existingActor.Img))
                 {
-                    System.IO.File.Delete(oldFilePath);
+                    var oldFilePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\img\\actors_images", existingActor.Img);
+                    if (System.IO.File.Exists(oldFilePath))
+                    {
+                        System.IO.File.Delete(oldFilePath);
+                    }
                 }
-                actor.Img = newFileName;
+                actorVM.Img = newFileName;
             }
-            else
-            {
-                actor.Img = existingActor.Img;
-            }
-            _context.Actors.Update(actor);
+            existingActor.Name = actorVM.Name;
+                existingActor.Description = actorVM.Description;
+                existingActor.MovieId = actorVM.MovieId;
+           
+           // _context.Actors.Update(actors);
             _context.SaveChanges();
             return RedirectToAction(nameof(Index));
         }
